+++ GHEtool/Borefield.py
@@ -2380,17 +2380,8 @@ class Borefield(BaseClass):
         TypeError
             Raises TypeError when a constant borehole thermal resistance is used.
         """
-        if self.borehole.use_constant_Rb:
-            raise TypeError("The inlet and outlet temperatures cannot be calculated when a constant effective borehole"
-                            "thermal resistance is used.")
-
-        delta_temp = power / (
-                self.borehole.fluid_data.cp(temperature=temperature) / 1000 *
-                self.borehole.flow_data.mfr_borefield(fluid_data=self.fluid_data, temperature=temperature,
-                                                      nb_of_boreholes=self.number_of_boreholes, power=power, **kwargs))
-        delta_temp = np.nan_to_num(delta_temp, )
-        # power < 0 when in extraction
-        return temperature + delta_temp / 2, temperature - delta_temp / 2
+        return self.borehole._calculate_borefield_inlet_outlet_temperature(power, temperature, self.number_of_boreholes,
+                                                                           **kwargs)
 
     def __export__(self):
         return {